Remove an unneeded type cast. (dm)

+++ Programs/usb_cp2101.c
@@ -79,7 +79,7 @@ usbVerifyBaudRate_CP2101 (UsbDevice *device, USB_CP2101_BaudRate expected) {
     logMessage(LOG_WARNING,
                "unexpected CP2101 baud rate value:"
                " Expected:%"PRIu32 " Actual:%"PRIu32,
-               (uint32_t)expected, getLittleEndian32(actual));
+               expected, getLittleEndian32(actual));
   } else {
     return 1;
   }
